Refactor Barcode AI Module into Barcode Scanner
- Removed BarcodeAiSettingsView and its associated functionality.
- Introduced BarcodeScannerRootView to handle barcode scanning features.
- Updated HomeRootView to present the new barcode scanner instead of the AI settings.
- Adjusted MainStateModel to route to the new barcode scanner.
- Removed AI settings link from ServicesView.
- Updated Screen enum to replace barcodeAiView with barcodeScanner and added barcodeScannerTreatment case.
