+++ Trio/Sources/Modules/BarcodeAi/BarcodeAiProvider.swift
@@ -1,0 +1,3 @@
+extension BarcodeScanner {
+    final class Provider: BaseProvider, BarcodeScannerProvider {}
+}
