--- Trio/Sources/Modules/BarcodeAi/BarcodeAiProvider.swift.swift
@@ -1,3 +1,0 @@
-extension BarcodeAi {
-    final class Provider: BaseProvider, BarcodeAiProvider {}
-}
