+++ Trio/Sources/Modules/Home/View/HomeRootView.swift
@@ -1200,7 +1200,7 @@ extension Home {
                 )
                 .simultaneousGesture(LongPressGesture().onEnded { _ in
                     didLongPress = true
-                    state.showModal(for: .barcodeAiView)
+                    state.showModal(for: .barcodeScanner)
                     playHaptics(haptics)
                     print("Secret Long Press Action!")
                 })
